% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_genotypes.R
\name{get_all_genotypes}
\alias{get_all_genotypes}
\title{Make a geno_full}
\usage{
get_all_genotypes(variants_df, pgen_list, directory)
}
\arguments{
\item{variants_df}{dataframe containing all of the variants in the directory, created with get_directory_variants}
}
\description{
Make a geno_full
}
\keyword{all}
\keyword{genotypes}
\keyword{get}
