% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chromosome_variants.R
\name{get_chromosome_variants}
\alias{get_chromosome_variants}
\title{Make a variant_df}
\usage{
get_chromosome_variants(chromosome, variants, directory)
}
\value{
variant_df
}
\description{
Make a variant_df
}
