% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genotypes_from_pgen.R
\name{get_genotypes_from_pgen}
\alias{get_genotypes_from_pgen}
\title{A tidyUkBioBank function
Function output: genotypes of all individuals of the variants of interest}
\usage{
get_genotypes_from_pgen(variants, variants_df, pgen_list)
}
\arguments{
\item{variants}{variants of interest}

\item{variants_df}{dataframe containing all of the variants in the directory, created with get_directory_variants}
}
\description{
A tidyUkBioBank function
Function output: genotypes of all individuals of the variants of interest
}
\examples{
get_genotypes_from_pgen()
}
\keyword{from}
\keyword{genotypes}
\keyword{get}
\keyword{pgen}
