% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variant_index.R
\name{get_variant_index}
\alias{get_variant_index}
\title{Make a variant index}
\usage{
get_variant_index(variant_id, pvar_handle, chromosome)
}
\value{
variant_index
}
\description{
Make a variant index
}
