% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_psam.R
\name{make_psam}
\alias{make_psam}
\title{Make a psam}
\usage{
make_psam(chromosome, directory)
}
\arguments{
\item{chromosome}{chromosome labeled as "chrNumber" (ex: "chr1")}

\item{directory}{directory containing the files}
}
\value{
a psam
}
\description{
Make a psam
}
