% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pvar.R
\name{make_pvar}
\alias{make_pvar}
\title{make pvar}
\usage{
make_pvar(chromosome, directory)
}
\arguments{
\item{chromosome}{which chromosome you are selecting and how it is labeled (eg: for chromosome 2 of sample3_chr2.pvar, chromsome = "chr2" OR for s1_c2.pvar, chromosome = "c2")}

\item{directory}{the directory with your pvar files}
}
\description{
make pvar
}
\keyword{pvar}
